backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Session
Flask-Cors
Werkzeug
redis
celery
python-dotenv
